refactor MoexDerivativeNamingHelper
